backend/requirements.txt
fastapi
uvicorn[standard]
python-binance
sqlmodel
python-dotenv
httpx
pydantic
psycopg2-binary
gunicorn
